+++ backend/hr/src/main/java/com/bbap/hr/repository/EmployeeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface EmployeeRepository extends JpaRepository<EmployeeEntity, Long> {
+public interface EmployeeRepository extends JpaRepository<EmployeeEntity, Integer> {
 
     Optional<EmployeeEntity> findByEmpNo(String empNo);
 }
